refactor(compilation): #1132 shared streaming-crypto type emitters
Extract RuntimeEmitter.StreamingCrypto.cs with the shared emitters for the
$Sign/$Verify and $Cipher/$Decipher pairs:

- Sign/Verify: the 100%-duplicated type definition (fields, algorithm-
  parsing ctor, UTF-8 Update) collapses into
  EmitStreamingSignVerifyTypeDefinition, parameterized only by type name
  and the two message strings. The two-phase split is preserved: phase 2
  (EmitTSSignFinalize/EmitTSVerifyFinalize) still emits Sign()/Verify()
  after runtime.SignDataBytes/VerifyDataBytes exist.
- Cipher/Decipher: shared EmitStreamingCipherCtor (direction flag picks
  CreateEncryptor/CreateDecryptor; Decipher inits its two buffer lists),
  EmitGcmTransformHelper (AesGcm.Encrypt vs Decrypt with its different
  parameter order), EmitCipherInputToBytes, EmitCipherBufferArgToBytes,
  and the SetAutoPadding/SetAAD/SetAuthTag/Dispose members.
  EmitCipherFormatOutput takes supportUtf8 — the deliberate #1054
  divergence stands: $Cipher emits hex/base64/buffer only, $Decipher
  additionally decodes utf8/utf-8.
- Hash/Hmac are structurally different (MemoryStream one-shot vs
  IncrementalHash) and are deliberately NOT unified; they now share only
  the EmitThrowIfFinalized guard.

Update/Final bodies stay per-type (different buffering targets and
transform flows) but route through the shared conversion helpers.
Emitted type/member names, EmittedRuntime ctor fields, and registration
order are unchanged.

Part of #1092, closes #1132
