+++ Compilation/RuntimeEmitter.TSCipher.cs
@@ -73,296 +73,22 @@ public partial class RuntimeEmitter
     /// </summary>
     private void EmitTSCipherCtor(TypeBuilder typeBuilder, EmittedRuntime runtime)
     {
-        var ctor = typeBuilder.DefineConstructor(
-            MethodAttributes.Public,
-            CallingConventions.Standard,
-            [_types.String, _types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte)]
-        );
-        runtime.TSCipherCtor = ctor;
-
-        var il = ctor.GetILGenerator();
-
-        // Locals
-        var lowerAlgoLocal = il.DeclareLocal(_types.String);
-        var isGcmLocal = il.DeclareLocal(_types.Boolean);
-        var keySizeLocal = il.DeclareLocal(_types.Int32);
-
-        // Call base constructor
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Call, _types.GetDefaultConstructor(_types.Object));
-
-        // _algorithm = algorithm.ToLowerInvariant()
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldarg_1);
-        il.Emit(OpCodes.Callvirt, _types.String.GetMethod("ToLowerInvariant")!);
-        il.Emit(OpCodes.Dup);
-        il.Emit(OpCodes.Stloc, lowerAlgoLocal);
-        il.Emit(OpCodes.Stfld, _tsCipherAlgorithmField);
-
-        // Parse algorithm - determine isGcm and keySize
-        var aes128CbcLabel = il.DefineLabel();
-        var aes192CbcLabel = il.DefineLabel();
-        var aes256CbcLabel = il.DefineLabel();
-        var aes128GcmLabel = il.DefineLabel();
-        var aes192GcmLabel = il.DefineLabel();
-        var aes256GcmLabel = il.DefineLabel();
-        var afterParseLabel = il.DefineLabel();
-        var unsupportedLabel = il.DefineLabel();
-
-        // Check algorithms
-        EmitStringCompare(il, lowerAlgoLocal, "aes-128-cbc", aes128CbcLabel);
-        EmitStringCompare(il, lowerAlgoLocal, "aes-192-cbc", aes192CbcLabel);
-        EmitStringCompare(il, lowerAlgoLocal, "aes-256-cbc", aes256CbcLabel);
-        EmitStringCompare(il, lowerAlgoLocal, "aes-128-gcm", aes128GcmLabel);
-        EmitStringCompare(il, lowerAlgoLocal, "aes-192-gcm", aes192GcmLabel);
-        EmitStringCompare(il, lowerAlgoLocal, "aes-256-gcm", aes256GcmLabel);
-        il.Emit(OpCodes.Br, unsupportedLabel);
-
-        // aes-128-cbc: keySize=16, isGcm=false
-        il.MarkLabel(aes128CbcLabel);
-        il.Emit(OpCodes.Ldc_I4, 16);
-        il.Emit(OpCodes.Stloc, keySizeLocal);
-        il.Emit(OpCodes.Ldc_I4_0);
-        il.Emit(OpCodes.Stloc, isGcmLocal);
-        il.Emit(OpCodes.Br, afterParseLabel);
-
-        // aes-192-cbc: keySize=24, isGcm=false
-        il.MarkLabel(aes192CbcLabel);
-        il.Emit(OpCodes.Ldc_I4, 24);
-        il.Emit(OpCodes.Stloc, keySizeLocal);
-        il.Emit(OpCodes.Ldc_I4_0);
-        il.Emit(OpCodes.Stloc, isGcmLocal);
-        il.Emit(OpCodes.Br, afterParseLabel);
-
-        // aes-256-cbc: keySize=32, isGcm=false
-        il.MarkLabel(aes256CbcLabel);
-        il.Emit(OpCodes.Ldc_I4, 32);
-        il.Emit(OpCodes.Stloc, keySizeLocal);
-        il.Emit(OpCodes.Ldc_I4_0);
-        il.Emit(OpCodes.Stloc, isGcmLocal);
-        il.Emit(OpCodes.Br, afterParseLabel);
-
-        // aes-128-gcm: keySize=16, isGcm=true
-        il.MarkLabel(aes128GcmLabel);
-        il.Emit(OpCodes.Ldc_I4, 16);
-        il.Emit(OpCodes.Stloc, keySizeLocal);
-        il.Emit(OpCodes.Ldc_I4_1);
-        il.Emit(OpCodes.Stloc, isGcmLocal);
-        il.Emit(OpCodes.Br, afterParseLabel);
-
-        // aes-192-gcm: keySize=24, isGcm=true
-        il.MarkLabel(aes192GcmLabel);
-        il.Emit(OpCodes.Ldc_I4, 24);
-        il.Emit(OpCodes.Stloc, keySizeLocal);
-        il.Emit(OpCodes.Ldc_I4_1);
-        il.Emit(OpCodes.Stloc, isGcmLocal);
-        il.Emit(OpCodes.Br, afterParseLabel);
-
-        // aes-256-gcm: keySize=32, isGcm=true
-        il.MarkLabel(aes256GcmLabel);
-        il.Emit(OpCodes.Ldc_I4, 32);
-        il.Emit(OpCodes.Stloc, keySizeLocal);
-        il.Emit(OpCodes.Ldc_I4_1);
-        il.Emit(OpCodes.Stloc, isGcmLocal);
-        il.Emit(OpCodes.Br, afterParseLabel);
-
-        // Unsupported algorithm - throw
-        il.MarkLabel(unsupportedLabel);
-        il.Emit(OpCodes.Ldstr, "Unsupported cipher algorithm: ");
-        il.Emit(OpCodes.Ldarg_1);
-        il.Emit(OpCodes.Call, _types.String.GetMethod("Concat", [_types.String, _types.String])!);
-        il.Emit(OpCodes.Newobj, _types.ArgumentException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        il.MarkLabel(afterParseLabel);
-
-        // Store _isGcm
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldloc, isGcmLocal);
-        il.Emit(OpCodes.Stfld, _tsCipherIsGcmField);
-
-        // Validate key size
-        var keySizeOkLabel = il.DefineLabel();
-        il.Emit(OpCodes.Ldarg_2);
-        il.Emit(OpCodes.Ldlen);
-        il.Emit(OpCodes.Conv_I4);
-        il.Emit(OpCodes.Ldloc, keySizeLocal);
-        il.Emit(OpCodes.Beq, keySizeOkLabel);
-
-        // Key size mismatch - throw
-        il.Emit(OpCodes.Ldstr, "Invalid key length for cipher algorithm");
-        il.Emit(OpCodes.Newobj, _types.ArgumentException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        il.MarkLabel(keySizeOkLabel);
-
-        // Validate IV size: GCM=12, CBC=16
-        var ivSizeOkLabel = il.DefineLabel();
-        var expectedIvSizeLocal = il.DeclareLocal(_types.Int32);
-
-        // expectedIvSize = isGcm ? 12 : 16
-        il.Emit(OpCodes.Ldloc, isGcmLocal);
-        var notGcmIvLabel = il.DefineLabel();
-        var storeExpectedIvLabel = il.DefineLabel();
-        il.Emit(OpCodes.Brfalse, notGcmIvLabel);
-        il.Emit(OpCodes.Ldc_I4, 12);
-        il.Emit(OpCodes.Br, storeExpectedIvLabel);
-        il.MarkLabel(notGcmIvLabel);
-        il.Emit(OpCodes.Ldc_I4, 16);
-        il.MarkLabel(storeExpectedIvLabel);
-        il.Emit(OpCodes.Stloc, expectedIvSizeLocal);
-
-        il.Emit(OpCodes.Ldarg_3);
-        il.Emit(OpCodes.Ldlen);
-        il.Emit(OpCodes.Conv_I4);
-        il.Emit(OpCodes.Ldloc, expectedIvSizeLocal);
-        il.Emit(OpCodes.Beq, ivSizeOkLabel);
-
-        // IV size mismatch - throw
-        il.Emit(OpCodes.Ldstr, "Invalid IV length for cipher algorithm");
-        il.Emit(OpCodes.Newobj, _types.ArgumentException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        il.MarkLabel(ivSizeOkLabel);
-
-        // Store key and iv
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldarg_2);
-        il.Emit(OpCodes.Stfld, _tsCipherKeyField);
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldarg_3);
-        il.Emit(OpCodes.Stfld, _tsCipherIvField);
-
-        // Initialize _plaintextBuffer for GCM
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Newobj, _types.GetDefaultConstructor(_types.ListOfByte));
-        il.Emit(OpCodes.Stfld, _tsCipherPlaintextBufferField);
-
-        // _autoPadding = true
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldc_I4_1);
-        il.Emit(OpCodes.Stfld, _tsCipherAutoPaddingField);
-
-        // _finalized = false
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldc_I4_0);
-        il.Emit(OpCodes.Stfld, _tsCipherFinalizedField);
-
-        // Initialize crypto objects based on mode
-        var initCbcLabel = il.DefineLabel();
-        var initDoneLabel = il.DefineLabel();
-
-        il.Emit(OpCodes.Ldloc, isGcmLocal);
-        il.Emit(OpCodes.Brfalse, initCbcLabel);
-
-        // GCM mode: _aesGcm = new AesGcm(_key, 16)
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldarg_2); // key
-        il.Emit(OpCodes.Ldc_I4, 16); // tag size
-        il.Emit(OpCodes.Newobj, _types.AesGcm.GetConstructor([_types.MakeArrayType(_types.Byte), _types.Int32])!);
-        il.Emit(OpCodes.Stfld, _tsCipherAesGcmField);
-        il.Emit(OpCodes.Br, initDoneLabel);
-
-        // CBC mode: create Aes and encryptor
-        il.MarkLabel(initCbcLabel);
-
-        // _aes = Aes.Create()
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Call, _types.Aes.GetMethod("Create", Type.EmptyTypes)!);
-        il.Emit(OpCodes.Stfld, _tsCipherAesField);
-
-        // _aes.Key = key
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Ldarg_2);
-        il.Emit(OpCodes.Callvirt, _types.Aes.GetProperty("Key")!.SetMethod!);
-
-        // _aes.IV = iv
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Ldarg_3);
-        il.Emit(OpCodes.Callvirt, _types.Aes.GetProperty("IV")!.SetMethod!);
-
-        // _aes.Mode = CipherMode.CBC
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Ldc_I4_1); // CipherMode.CBC = 1
-        il.Emit(OpCodes.Callvirt, _types.Aes.GetProperty("Mode")!.SetMethod!);
-
-        // _aes.Padding = PaddingMode.PKCS7
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Ldc_I4_2); // PaddingMode.PKCS7 = 2
-        il.Emit(OpCodes.Callvirt, _types.Aes.GetProperty("Padding")!.SetMethod!);
-
-        // _encryptor = _aes.CreateEncryptor()
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Callvirt, _types.Aes.GetMethod("CreateEncryptor", Type.EmptyTypes)!);
-        il.Emit(OpCodes.Stfld, _tsCipherEncryptorField);
-
-        il.MarkLabel(initDoneLabel);
-        il.Emit(OpCodes.Ret);
+        runtime.TSCipherCtor = EmitStreamingCipherCtor(typeBuilder, CipherFields(), isEncrypt: true,
+            [_tsCipherPlaintextBufferField]);
     }
 
-    private void EmitStringCompare(ILGenerator il, LocalBuilder stringLocal, string value, Label targetLabel)
-    {
-        il.Emit(OpCodes.Ldloc, stringLocal);
-        il.Emit(OpCodes.Ldstr, value);
-        il.Emit(OpCodes.Call, _types.String.GetMethod("op_Equality", [_types.String, _types.String])!);
-        il.Emit(OpCodes.Brtrue, targetLabel);
-    }
+    private StreamingCipherFields CipherFields() => new(
+        _tsCipherAlgorithmField, _tsCipherKeyField, _tsCipherIvField, _tsCipherIsGcmField,
+        _tsCipherAesField, _tsCipherEncryptorField, _tsCipherAesGcmField,
+        _tsCipherFinalizedField, _tsCipherAutoPaddingField, _tsCipherAuthTagField, _tsCipherAadField);
 
     /// <summary>
     /// Emits: private static void GcmEncryptHelper(AesGcm gcm, byte[] nonce, byte[] plaintext, byte[] ciphertext, byte[] tag, byte[] aad)
-    /// Uses the AesGcm.Encrypt byte[] overload.
+    /// via the shared <see cref="EmitGcmTransformHelper"/>.
     /// </summary>
     private void EmitTSCipherGcmEncryptHelper(TypeBuilder typeBuilder, EmittedRuntime runtime)
     {
-        var method = typeBuilder.DefineMethod(
-            "GcmEncryptHelper",
-            MethodAttributes.Private | MethodAttributes.Static,
-            _types.Void,
-            [_types.AesGcm, _types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte),
-             _types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte)]);
-        _tsCipherGcmEncryptHelper = method;
-
-        var il = method.GetILGenerator();
-
-        // Get the byte[] overload: Encrypt(byte[] nonce, byte[] plaintext, byte[] ciphertext, byte[] tag, byte[] associatedData)
-        var encryptMethod = _types.AesGcm.GetMethod("Encrypt",
-            [_types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte),
-             _types.MakeArrayType(_types.Byte), _types.MakeArrayType(_types.Byte)])!;
-
-        // Check if aad is null - if so, use empty array
-        var aadNotNullLabel = il.DefineLabel();
-        var callEncryptLabel = il.DefineLabel();
-        var aadLocal = il.DeclareLocal(_types.MakeArrayType(_types.Byte));
-
-        il.Emit(OpCodes.Ldarg_S, (byte)5);
-        il.Emit(OpCodes.Brtrue, aadNotNullLabel);
-
-        // aad is null - use empty byte array
-        il.Emit(OpCodes.Ldc_I4_0);
-        il.Emit(OpCodes.Newarr, _types.Byte);
-        il.Emit(OpCodes.Stloc, aadLocal);
-        il.Emit(OpCodes.Br, callEncryptLabel);
-
-        il.MarkLabel(aadNotNullLabel);
-        il.Emit(OpCodes.Ldarg_S, (byte)5);
-        il.Emit(OpCodes.Stloc, aadLocal);
-
-        // Call gcm.Encrypt(nonce, plaintext, ciphertext, tag, aad)
-        il.MarkLabel(callEncryptLabel);
-        il.Emit(OpCodes.Ldarg_0);   // gcm
-        il.Emit(OpCodes.Ldarg_1);   // nonce
-        il.Emit(OpCodes.Ldarg_2);   // plaintext
-        il.Emit(OpCodes.Ldarg_3);   // ciphertext
-        il.Emit(OpCodes.Ldarg_S, (byte)4);   // tag
-        il.Emit(OpCodes.Ldloc, aadLocal);    // aad
-        il.Emit(OpCodes.Callvirt, encryptMethod);
-        il.Emit(OpCodes.Ret);
+        _tsCipherGcmEncryptHelper = EmitGcmTransformHelper(typeBuilder, "GcmEncryptHelper", isEncrypt: true);
     }
 
     /// <summary>
@@ -382,20 +108,11 @@ public partial class RuntimeEmitter
 
         var il = method.GetILGenerator();
 
-        // Check if finalized
-        var notFinalizedLabel = il.DefineLabel();
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherFinalizedField);
-        il.Emit(OpCodes.Brfalse, notFinalizedLabel);
-        il.Emit(OpCodes.Ldstr, "Cipher has already been finalized");
-        il.Emit(OpCodes.Newobj, _types.InvalidOperationException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        il.MarkLabel(notFinalizedLabel);
+        EmitThrowIfFinalized(il, _tsCipherFinalizedField, "Cipher has already been finalized");
 
         // Convert input to bytes
         var inputBytesLocal = il.DeclareLocal(_types.MakeArrayType(_types.Byte));
-        EmitConvertToBytes(il, runtime, OpCodes.Ldarg_1, OpCodes.Ldarg_2);
+        EmitCipherInputToBytes(il, runtime, OpCodes.Ldarg_1, OpCodes.Ldarg_2);
         il.Emit(OpCodes.Stloc, inputBytesLocal);
 
         // Buffer data in _plaintextBuffer (used for both CBC and GCM)
@@ -410,7 +127,7 @@ public partial class RuntimeEmitter
         il.Emit(OpCodes.Ldc_I4_0);
         il.Emit(OpCodes.Newarr, _types.Byte);
         il.Emit(OpCodes.Stloc, resultLocal);
-        EmitFormatOutput(il, runtime, resultLocal, OpCodes.Ldarg_3);
+        EmitCipherFormatOutput(il, runtime, resultLocal, OpCodes.Ldarg_3, supportUtf8: false);
         il.Emit(OpCodes.Ret);
     }
 
@@ -429,16 +146,7 @@ public partial class RuntimeEmitter
 
         var il = method.GetILGenerator();
 
-        // Check if finalized
-        var notFinalizedLabel = il.DefineLabel();
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherFinalizedField);
-        il.Emit(OpCodes.Brfalse, notFinalizedLabel);
-        il.Emit(OpCodes.Ldstr, "Cipher has already been finalized");
-        il.Emit(OpCodes.Newobj, _types.InvalidOperationException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        il.MarkLabel(notFinalizedLabel);
+        EmitThrowIfFinalized(il, _tsCipherFinalizedField, "Cipher has already been finalized");
 
         // Set finalized
         il.Emit(OpCodes.Ldarg_0);
@@ -491,7 +199,7 @@ public partial class RuntimeEmitter
         il.Emit(OpCodes.Stfld, _tsCipherAuthTagField);
 
         // Return formatted ciphertext
-        EmitFormatOutput(il, runtime, ciphertextLocal, OpCodes.Ldarg_1);
+        EmitCipherFormatOutput(il, runtime, ciphertextLocal, OpCodes.Ldarg_1, supportUtf8: false);
         il.Emit(OpCodes.Ret);
 
         // CBC mode: Get all buffered data and TransformFinalBlock
@@ -516,7 +224,7 @@ public partial class RuntimeEmitter
         il.Emit(OpCodes.Callvirt, _types.ICryptoTransform.GetMethod("TransformFinalBlock")!);
         il.Emit(OpCodes.Stloc, finalBlockLocal);
 
-        EmitFormatOutput(il, runtime, finalBlockLocal, OpCodes.Ldarg_1);
+        EmitCipherFormatOutput(il, runtime, finalBlockLocal, OpCodes.Ldarg_1, supportUtf8: false);
         il.Emit(OpCodes.Ret);
     }
 
@@ -524,21 +232,7 @@ public partial class RuntimeEmitter
     /// Emits: public $Cipher SetAutoPadding(bool autoPadding)
     /// </summary>
     private void EmitTSCipherSetAutoPadding(TypeBuilder typeBuilder, EmittedRuntime runtime)
-    {
-        var method = typeBuilder.DefineMethod(
-            "SetAutoPadding",
-            MethodAttributes.Public,
-            typeBuilder,
-            [_types.Boolean]
-        );
-        _ = method;
-
-        var il = method.GetILGenerator();
-
-        // Return this for chaining (simplified - just return this)
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ret);
-    }
+        => EmitCipherSetAutoPadding(typeBuilder);
 
     /// <summary>
     /// Emits: public object GetAuthTag()
@@ -585,294 +279,15 @@ public partial class RuntimeEmitter
     }
 
     /// <summary>
-    /// Emits: public $Cipher SetAAD(object aad)
-    /// Accepts $Buffer or byte[] and stores as byte[].
+    /// Emits: public $Cipher SetAAD(object aad) — accepts $Buffer or byte[].
     /// </summary>
     private void EmitTSCipherSetAAD(TypeBuilder typeBuilder, EmittedRuntime runtime)
-    {
-        var method = typeBuilder.DefineMethod(
-            "SetAAD",
-            MethodAttributes.Public,
-            typeBuilder,
-            [_types.Object]
-        );
-        _ = method;
-
-        var il = method.GetILGenerator();
-
-        // Convert $Buffer to byte[]
-        var bytesLocal = il.DeclareLocal(_types.MakeArrayType(_types.Byte));
-        EmitCipherBufferToBytes(il, runtime, OpCodes.Ldarg_1);
-        il.Emit(OpCodes.Stloc, bytesLocal);
-
-        // Store AAD
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldloc, bytesLocal);
-        il.Emit(OpCodes.Stfld, _tsCipherAadField);
-
-        // Return this
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ret);
-    }
+        => EmitCipherStoreBytesArg(typeBuilder, runtime, "SetAAD", _tsCipherAadField);
 
     /// <summary>
     /// Emits: public void Dispose()
     /// </summary>
     private void EmitTSCipherDispose(TypeBuilder typeBuilder, EmittedRuntime runtime)
-    {
-        var method = typeBuilder.DefineMethod(
-            "Dispose",
-            MethodAttributes.Public | MethodAttributes.Virtual,
-            _types.Void,
-            Type.EmptyTypes
-        );
+        => EmitCipherDispose(typeBuilder, _tsCipherEncryptorField, _tsCipherAesField, _tsCipherAesGcmField);
 
-        var il = method.GetILGenerator();
-
-        // Dispose encryptor if not null
-        var encryptorNullLabel = il.DefineLabel();
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherEncryptorField);
-        il.Emit(OpCodes.Brfalse, encryptorNullLabel);
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherEncryptorField);
-        il.Emit(OpCodes.Callvirt, _types.IDisposable.GetMethod("Dispose")!);
-        il.MarkLabel(encryptorNullLabel);
-
-        // Dispose aes if not null
-        var aesNullLabel = il.DefineLabel();
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Brfalse, aesNullLabel);
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesField);
-        il.Emit(OpCodes.Callvirt, _types.IDisposable.GetMethod("Dispose")!);
-        il.MarkLabel(aesNullLabel);
-
-        // Dispose aesGcm if not null
-        var aesGcmNullLabel = il.DefineLabel();
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesGcmField);
-        il.Emit(OpCodes.Brfalse, aesGcmNullLabel);
-        il.Emit(OpCodes.Ldarg_0);
-        il.Emit(OpCodes.Ldfld, _tsCipherAesGcmField);
-        il.Emit(OpCodes.Callvirt, _types.IDisposable.GetMethod("Dispose")!);
-        il.MarkLabel(aesGcmNullLabel);
-
-        il.Emit(OpCodes.Ret);
-    }
-
-    /// <summary>
-    /// Helper to emit code that converts input data to byte array.
-    /// Supports string with encoding (hex, base64, utf8) or Buffer.
-    /// </summary>
-    private void EmitConvertToBytes(ILGenerator il, EmittedRuntime runtime, OpCode loadData, OpCode loadEncoding)
-    {
-        var dataLocal = il.DeclareLocal(_types.Object);
-        var encodingLocal = il.DeclareLocal(_types.String);
-
-        il.Emit(loadData);
-        il.Emit(OpCodes.Stloc, dataLocal);
-
-        // Encoding may be object ($Undefined, null, or string) — extract string or null
-        var encodingIsStringLabel = il.DefineLabel();
-        var encodingResolvedLabel = il.DefineLabel();
-        il.Emit(loadEncoding);
-        il.Emit(OpCodes.Isinst, _types.String);
-        il.Emit(OpCodes.Dup);
-        il.Emit(OpCodes.Brtrue, encodingIsStringLabel);
-        il.Emit(OpCodes.Pop);
-        il.Emit(OpCodes.Ldnull);
-        il.MarkLabel(encodingIsStringLabel);
-        il.MarkLabel(encodingResolvedLabel);
-        il.Emit(OpCodes.Stloc, encodingLocal);
-
-        var isBufferLabel = il.DefineLabel();
-        var isStringLabel = il.DefineLabel();
-        var checkHexLabel = il.DefineLabel();
-        var checkBase64Label = il.DefineLabel();
-        var utf8DefaultLabel = il.DefineLabel();
-        var doneLabel = il.DefineLabel();
-
-        // Check if Buffer
-        il.Emit(OpCodes.Ldloc, dataLocal);
-        il.Emit(OpCodes.Isinst, runtime.TSBufferType);
-        il.Emit(OpCodes.Brtrue, isBufferLabel);
-
-        // Check if string
-        il.Emit(OpCodes.Ldloc, dataLocal);
-        il.Emit(OpCodes.Isinst, _types.String);
-        il.Emit(OpCodes.Brtrue, isStringLabel);
-
-        // Default: throw
-        il.Emit(OpCodes.Ldstr, "Data must be a string or Buffer");
-        il.Emit(OpCodes.Newobj, _types.ArgumentException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        // Buffer path
-        il.MarkLabel(isBufferLabel);
-        il.Emit(OpCodes.Ldloc, dataLocal);
-        il.Emit(OpCodes.Castclass, runtime.TSBufferType);
-        il.Emit(OpCodes.Call, runtime.TSBufferGetData);
-        il.Emit(OpCodes.Br, doneLabel);
-
-        // String path - check encoding
-        il.MarkLabel(isStringLabel);
-
-        // Check if encoding is null
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Brfalse, utf8DefaultLabel);
-
-        // lowerEncoding = encoding.ToLowerInvariant()
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Callvirt, _types.String.GetMethod("ToLowerInvariant")!);
-        il.Emit(OpCodes.Stloc, encodingLocal);
-
-        // Check "hex"
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Ldstr, "hex");
-        il.Emit(OpCodes.Call, _types.String.GetMethod("op_Equality", [_types.String, _types.String])!);
-        il.Emit(OpCodes.Brtrue, checkHexLabel);
-
-        // Check "base64"
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Ldstr, "base64");
-        il.Emit(OpCodes.Call, _types.String.GetMethod("op_Equality", [_types.String, _types.String])!);
-        il.Emit(OpCodes.Brtrue, checkBase64Label);
-
-        // Default UTF8
-        il.Emit(OpCodes.Br, utf8DefaultLabel);
-
-        // Hex decode
-        il.MarkLabel(checkHexLabel);
-        il.Emit(OpCodes.Ldloc, dataLocal);
-        il.Emit(OpCodes.Castclass, _types.String);
-        il.Emit(OpCodes.Call, _types.ConvertFromHexString);
-        il.Emit(OpCodes.Br, doneLabel);
-
-        // Base64 decode
-        il.MarkLabel(checkBase64Label);
-        il.Emit(OpCodes.Ldloc, dataLocal);
-        il.Emit(OpCodes.Castclass, _types.String);
-        il.Emit(OpCodes.Call, _types.ConvertFromBase64String);
-        il.Emit(OpCodes.Br, doneLabel);
-
-        // UTF8 default
-        il.MarkLabel(utf8DefaultLabel);
-        il.Emit(OpCodes.Call, _types.Encoding.GetProperty("UTF8")!.GetGetMethod()!);
-        il.Emit(OpCodes.Ldloc, dataLocal);
-        il.Emit(OpCodes.Castclass, _types.String);
-        il.Emit(OpCodes.Callvirt, _types.Encoding.GetMethod("GetBytes", [_types.String])!);
-
-        il.MarkLabel(doneLabel);
-    }
-
-    /// <summary>
-    /// Helper to emit code that formats output bytes.
-    /// </summary>
-    private void EmitFormatOutput(ILGenerator il, EmittedRuntime runtime, LocalBuilder bytesLocal, OpCode loadEncoding)
-    {
-        // Check encoding
-        var checkHexLabel = il.DefineLabel();
-        var checkBase64Label = il.DefineLabel();
-        var returnBufferLabel = il.DefineLabel();
-        var encodingLocal = il.DeclareLocal(_types.String);
-
-        // Encoding may be object ($Undefined, null, or string) — extract string or null
-        var outEncIsStringLabel = il.DefineLabel();
-        il.Emit(loadEncoding);
-        il.Emit(OpCodes.Isinst, _types.String);
-        il.Emit(OpCodes.Dup);
-        il.Emit(OpCodes.Brtrue, outEncIsStringLabel);
-        il.Emit(OpCodes.Pop);
-        il.Emit(OpCodes.Ldnull);
-        il.MarkLabel(outEncIsStringLabel);
-        il.Emit(OpCodes.Stloc, encodingLocal);
-
-        // if (encoding == null) return Buffer
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Brfalse, returnBufferLabel);
-
-        // lowerEncoding = encoding.ToLowerInvariant()
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Callvirt, _types.String.GetMethod("ToLowerInvariant")!);
-        il.Emit(OpCodes.Stloc, encodingLocal);
-
-        // Check "hex"
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Ldstr, "hex");
-        il.Emit(OpCodes.Call, _types.String.GetMethod("op_Equality", [_types.String, _types.String])!);
-        il.Emit(OpCodes.Brtrue, checkHexLabel);
-
-        // Check "base64"
-        il.Emit(OpCodes.Ldloc, encodingLocal);
-        il.Emit(OpCodes.Ldstr, "base64");
-        il.Emit(OpCodes.Call, _types.String.GetMethod("op_Equality", [_types.String, _types.String])!);
-        il.Emit(OpCodes.Brtrue, checkBase64Label);
-
-        // Default - return buffer
-        il.Emit(OpCodes.Br, returnBufferLabel);
-
-        // Return hex string
-        il.MarkLabel(checkHexLabel);
-        il.Emit(OpCodes.Ldloc, bytesLocal);
-        il.Emit(OpCodes.Call, _types.ConvertToHexString);
-        il.Emit(OpCodes.Callvirt, _types.String.GetMethod("ToLowerInvariant")!);
-        il.Emit(OpCodes.Ret);
-
-        // Return base64 string
-        il.MarkLabel(checkBase64Label);
-        il.Emit(OpCodes.Ldloc, bytesLocal);
-        il.Emit(OpCodes.Call, _types.ConvertToBase64String);
-        il.Emit(OpCodes.Ret);
-
-        // Return Buffer
-        il.MarkLabel(returnBufferLabel);
-        il.Emit(OpCodes.Ldloc, bytesLocal);
-        il.Emit(OpCodes.Newobj, runtime.TSBufferCtor);
-    }
-
-    /// <summary>
-    /// Helper to emit code that converts a $Buffer or byte[] to byte[].
-    /// Leaves byte[] on stack.
-    /// </summary>
-    private void EmitCipherBufferToBytes(ILGenerator il, EmittedRuntime runtime, OpCode loadData)
-    {
-        var objLocal = il.DeclareLocal(_types.Object);
-        il.Emit(loadData);
-        il.Emit(OpCodes.Stloc, objLocal);
-
-        var isBufferLabel = il.DefineLabel();
-        var isByteArrayLabel = il.DefineLabel();
-        var doneLabel = il.DefineLabel();
-
-        // Check if $Buffer
-        il.Emit(OpCodes.Ldloc, objLocal);
-        il.Emit(OpCodes.Isinst, runtime.TSBufferType);
-        il.Emit(OpCodes.Brtrue, isBufferLabel);
-
-        // Check if byte[]
-        il.Emit(OpCodes.Ldloc, objLocal);
-        il.Emit(OpCodes.Isinst, _types.MakeArrayType(_types.Byte));
-        il.Emit(OpCodes.Brtrue, isByteArrayLabel);
-
-        // Fallback: throw
-        il.Emit(OpCodes.Ldstr, "Expected Buffer or byte[]");
-        il.Emit(OpCodes.Newobj, _types.ArgumentException.GetConstructor([_types.String])!);
-        il.Emit(OpCodes.Throw);
-
-        // Buffer path: call GetData()
-        il.MarkLabel(isBufferLabel);
-        il.Emit(OpCodes.Ldloc, objLocal);
-        il.Emit(OpCodes.Castclass, runtime.TSBufferType);
-        il.Emit(OpCodes.Call, runtime.TSBufferGetData);
-        il.Emit(OpCodes.Br, doneLabel);
-
-        // byte[] path: just cast
-        il.MarkLabel(isByteArrayLabel);
-        il.Emit(OpCodes.Ldloc, objLocal);
-        il.Emit(OpCodes.Castclass, _types.MakeArrayType(_types.Byte));
-
-        il.MarkLabel(doneLabel);
-    }
 }
